# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusOS Backend",
    "description": "\n    CampusOS Backend API \u2014 Trusted Operating System for African Universities.\n    Powered by Quai Network & Blip Pay.\n    \n    ### Flagship Module: Verified Student Identity (Milestone 2 & Milestone 3)\n    - Multi-document upload to Cloudinary (Student ID, Admission Letter)\n    - Institutional email (.edu.ng) validation\n    - Administrative verification queue & review workflow\n    - SHA-256 cryptographic credential hashing\n    - Quai Network on-chain credential registration (`StudentIdentity` smart contract)\n    - Permanent cryptographically signed Campus Identity QR\n    - Strict OWASP security headers, rate limiting, and magic-bytes file validation\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/wallet/connect": {
      "post": {
        "tags": [
          "Campus Wallet (Quai Network)"
        ],
        "summary": "Connect and authenticate Quai EVM wallet via signed message",
        "description": "Validates Web3 cryptographic signature to bind an EVM wallet address to a CampusOS student account.",
        "operationId": "connect_wallet_api_v1_wallet_connect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletConnectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/balance": {
      "get": {
        "tags": [
          "Campus Wallet (Quai Network)"
        ],
        "summary": "Get real-time Quai Network balance and NGN equivalent",
        "description": "Fetches live QUAI testnet balance from JSON-RPC node and converts to NGN fiat valuation.",
        "operationId": "get_wallet_balance_api_v1_wallet_balance_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "UUID of the student (optional if wallet_address is provided)",
              "title": "User Id"
            },
            "description": "UUID of the student (optional if wallet_address is provided)"
          },
          {
            "name": "wallet_address",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Quai EVM wallet address (optional if user_id is provided)",
              "title": "Wallet Address"
            },
            "description": "Quai EVM wallet address (optional if user_id is provided)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletBalanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/history": {
      "get": {
        "tags": [
          "Campus Wallet (Quai Network)"
        ],
        "summary": "Get paginated Quai transaction history",
        "description": "Returns chronological audit trail of all P2P sends, receives, deposits, and faucet claims.",
        "operationId": "get_wallet_history_api_v1_wallet_history_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "UUID of the student",
              "title": "User Id"
            },
            "description": "UUID of the student"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "title": "Response Get Wallet History Api V1 Wallet History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/send": {
      "post": {
        "tags": [
          "Campus Wallet (Quai Network)"
        ],
        "summary": "Send QUAI to a student or Quai EVM address",
        "description": "Transfers QUAI testnet tokens to a recipient address, email, or student UUID and logs immutable receipt.",
        "operationId": "send_quai_api_v1_wallet_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/dashboard/{user_id}": {
      "get": {
        "tags": [
          "Campus Wallet (Quai Network)"
        ],
        "summary": "Get comprehensive Campus Wallet dashboard profile",
        "description": "Returns connected wallet, live balance, transaction history, QR receive address, and trust score.",
        "operationId": "get_wallet_dashboard_api_v1_wallet_dashboard__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "recipient_address": {
            "type": "string",
            "title": "Recipient Address"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "tx_hash": {
            "type": "string",
            "title": "Tx Hash"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "network": {
            "type": "string",
            "title": "Network"
          },
          "block_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Number"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "wallet_address",
          "recipient_address",
          "amount",
          "tx_hash",
          "type",
          "status",
          "network",
          "created_at"
        ],
        "title": "TransactionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WalletBalanceResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "wallet_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wallet Address"
          },
          "balance_quai": {
            "type": "number",
            "title": "Balance Quai"
          },
          "balance_wei": {
            "type": "string",
            "title": "Balance Wei"
          },
          "fiat_value_ngn": {
            "type": "number",
            "title": "Fiat Value Ngn"
          },
          "network": {
            "type": "string",
            "title": "Network",
            "default": "Quai Network Testnet"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "balance_quai",
          "balance_wei",
          "fiat_value_ngn"
        ],
        "title": "WalletBalanceResponse"
      },
      "WalletConnectRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "UUID of the student"
          },
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address",
            "description": "Quai EVM wallet address (0x...)"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Signed cryptographic authentication challenge string"
          },
          "signature": {
            "type": "string",
            "title": "Signature",
            "description": "65-byte hex signature (0x...) from WalletConnect / Web3 provider"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "wallet_address",
          "message",
          "signature"
        ],
        "title": "WalletConnectRequest"
      },
      "WalletConnectResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "wallet_address",
          "verified",
          "message"
        ],
        "title": "WalletConnectResponse"
      },
      "WalletDashboardResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "wallet_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wallet Address"
          },
          "balance": {
            "$ref": "#/components/schemas/WalletBalanceResponse"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Transactions"
          },
          "qr_receive_address": {
            "type": "string",
            "title": "Qr Receive Address"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified"
          },
          "trust_score": {
            "type": "integer",
            "title": "Trust Score"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "balance",
          "transactions",
          "qr_receive_address",
          "is_verified",
          "trust_score"
        ],
        "title": "WalletDashboardResponse"
      },
      "WalletSendRequest": {
        "properties": {
          "sender_id": {
            "type": "string",
            "title": "Sender Id",
            "description": "UUID of the sender student"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient",
            "description": "Quai EVM address (0x...), email, or user UUID of the recipient"
          },
          "amount_quai": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount Quai",
            "description": "Amount of QUAI to transfer"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash",
            "description": "Optional transaction hash if signed and broadcasted by client wallet"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note",
            "description": "Optional transfer note or memo"
          }
        },
        "type": "object",
        "required": [
          "sender_id",
          "recipient",
          "amount_quai"
        ],
        "title": "WalletSendRequest"
      },
      "WalletSendResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tx_hash": {
            "type": "string",
            "title": "Tx Hash"
          },
          "amount_quai": {
            "type": "number",
            "title": "Amount Quai"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tx_hash",
          "amount_quai",
          "recipient",
          "status",
          "message"
        ],
        "title": "WalletSendResponse"
      }
    }
  }
}